backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
SpeechRecognition==3.10.0
pydub==0.25.1
gTTS==2.4.0
langdetect==1.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
